Updating functional tests to use port 9080
Due to the design of the functional tets and the requirements of the architecture we have to forward requests for the functional test port to the docker host (which is bound to the instance of mobi used for the functional tests). However, since we run other tests that do not use docker we must use a different port so the functional test port forwarding does not break the integration tests.